requirements.txt
2captcha-python==1.2.1
absl-py==1.4.0
accelerate==0.18.0
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.14.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bert-score==0.3.13
blanc==0.3.1
blis==0.7.9
boto3==1.28.2
botocore==1.31.2
bottle==0.12.25
cachetools==5.3.1
catalogue==2.0.8
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
colorcet==3.0.1
contourpy==1.1.0
crfm-helm==0.2.2
cycler==0.11.0
cymem==2.0.7
Cython==0.29.36
dacite==1.6.0
datasets==2.13.1
dill==0.3.5.1
emoji==2.6.0
exceptiongroup==1.1.2
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gdown==4.4.0
gin-config==0.5.0
google-api-core==2.11.1
google-api-python-client==2.64.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
icetk==0.0.4
idna==3.4
importlib-resources==5.10.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
kiwisolver==1.4.4
langcodes==3.3.0
llvmlite==0.39.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.6.3
moverscore==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.13
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
openai==0.27.8
outcome==1.2.0
packaging==23.1
pandas==2.0.3
param==1.13.0
parameterized==0.9.0
pathy==0.10.2
Pillow==10.0.0
portalocker==2.7.0
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pyct==0.5.0
pydantic==1.8.2
pyemd==0.5.1
pyext==0.7
pyhocon==0.3.60
pymongo==4.2.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytrec-eval==0.5
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
retrying==1.3.4
revChatGPT==0.1.1
rouge-score==0.1.2
rsa==4.9
s3transfer==0.6.1
sacrebleu==2.2.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.11.2
selenium==4.10.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlitedict==1.7.0
srsly==2.4.6
stanza==1.5.0
summ-eval==0.892
sympy==1.11.1
tabulate==0.9.0
thinc==8.0.17
threadpoolctl==3.1.0
tls-client==0.2.1
tokenizers==0.13.3
tqdm==4.64.1
transformers==4.30.2
trio==0.22.1
trio-websocket==0.10.3
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
uncertainty-calibration==0.1.4
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
wasabi==0.10.1
websocket-client==1.3.3
websockets==11.0.3
wsproto==1.2.0
xlrd==2.0.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.0
zstandard==0.18.0
